Skrur på aiven for identhendelser og sletter all ubrukt onprem oppsett
