+++ src/main/kotlin/no/nav/familie/baks/mottak/hendelser/IdenthendelseV2Consumer.kt
@@ -3,8 +3,10 @@ package no.nav.familie.baks.mottak.hendelser
 import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.Metrics
 import no.nav.familie.baks.mottak.integrasjoner.SakClient
+import no.nav.familie.kontrakter.felles.PersonIdent
 import no.nav.familie.log.mdc.MDCConstants
 import no.nav.person.pdl.aktor.v2.Aktor
+import no.nav.person.pdl.aktor.v2.Type
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -39,24 +41,23 @@ class IdenthendelseV2Consumer(
     fun listen(consumerRecord: ConsumerRecord<String, Aktor?>, ack: Acknowledgment) {
         try {
             Thread.sleep(60000) // Sender man med en gang, så får man Person ikke funnet fra PDL når ba-sak gjør filtrering. Venter derfor 1 minutt
-            MDC.put(MDCConstants.MDC_CALL_ID, UUID.randomUUID().toString() + "_a")
+            MDC.put(MDCConstants.MDC_CALL_ID, UUID.randomUUID().toString())
             SECURE_LOGGER.info("Har mottatt ident-hendelse $consumerRecord")
-            // TODO kommentert ut sending av identhendelser til testing er over og man er klar til å sende over
-//
-//            val aktør = consumerRecord.value()
-//
-//            if (aktør == null) {
-//                log.warn("Tom aktør fra identhendelse")
-//                SECURE_LOGGER.warn("Tom aktør fra identhendelse med nøkkel ${consumerRecord.key()}")
-//            }
-//
-//            aktør?.identifikatorer?.singleOrNull { ident ->
-//                ident.type == Type.FOLKEREGISTERIDENT && ident.gjeldende
-//            }?.also { folkeregisterident ->
-//                SECURE_LOGGER.info("Sender ident-hendelse til ba-sak for ident $folkeregisterident")
-//
-//                sakClient.sendIdenthendelseTilSak(PersonIdent(ident = folkeregisterident.idnummer.toString()))
-//            }
+
+            val aktør = consumerRecord.value()
+
+            if (aktør == null) {
+                log.warn("Tom aktør fra identhendelse")
+                SECURE_LOGGER.warn("Tom aktør fra identhendelse med nøkkel ${consumerRecord.key()}")
+            }
+
+            aktør?.identifikatorer?.singleOrNull { ident ->
+                ident.type == Type.FOLKEREGISTERIDENT && ident.gjeldende
+            }?.also { folkeregisterident ->
+                SECURE_LOGGER.info("Sender ident-hendelse til ba-sak for ident $folkeregisterident")
+
+                sakClient.sendIdenthendelseTilSak(PersonIdent(ident = folkeregisterident.idnummer.toString()))
+            }
         } catch (e: RuntimeException) {
             identhendelseFeiletCounter.increment()
             log.warn("Feil i prosessering av ident-hendelser", e)
