+++ src/main/resources/application-preprod.yaml
@@ -69,12 +69,6 @@ no.nav.security.jwt:
           client-jwk: ${TOKEN_X_PRIVATE_JWK}
 
 spring:
-  kafka:
-    bootstrap-servers: b27apvl00045.preprod.local:8443,b27apvl00046.preprod.local:8443,b27apvl00047.preprod.local:8443
-    properties:
-      schema.registry.url: https://kafka-schema-registry.nais-q.adeo.no
-    consumer:
-      auto-offset-reset: latest
   datasource:
     url: jdbc:postgresql://${DB_HOST}:${DB_PORT}/familie-baks-mottak
     username: ${DB_USERNAME}
