+++ src/main/resources/application-prod.yaml
@@ -69,10 +69,6 @@ no.nav.security.jwt:
           client-jwk: ${TOKEN_X_PRIVATE_JWK}
 
 spring:
-  kafka:
-    bootstrap-servers: a01apvl00145.adeo.no:8443,a01apvl00146.adeo.no:8443,a01apvl00147.adeo.no:8443,a01apvl00148.adeo.no:8443,a01apvl00149.adeo.no:8443,a01apvl00150.adeo.no:8443
-    properties:
-      schema.registry.url: https://kafka-schema-registry.nais.adeo.no
   datasource:
     url: jdbc:postgresql://${DB_HOST}:${DB_PORT}/familie-baks-mottak
     username: ${DB_USERNAME}
